feat (asv-avalonia): refactor properties
Separate properties into partial classes for TelemBox, WorkspacePanel, WorkspaceItem, and BatteryIndicator.
Asana: https://app.asana.com/0/1203851531040615/1209709314147235/f
