+++ src/Asv.Avalonia.Map/Controls/ItemsControl/MapItemsControl.axaml.cs
@@ -1,12 +1,7 @@
-﻿using System.Collections;
-using Asv.Common;
-using Avalonia;
+﻿using Avalonia;
 using Avalonia.Controls;
 using Avalonia.Controls.Primitives;
-using Avalonia.Controls.Selection;
-using Avalonia.Controls.Templates;
 using Avalonia.Input;
-using Avalonia.Metadata;
 
 namespace Asv.Avalonia.Map;
 
@@ -18,9 +13,9 @@ public enum DragState
     SelectRectangle,
 }
 
-public class MapItemsControl : SelectingItemsControl
+public partial class MapItemsControl : SelectingItemsControl
 {
-    private Point _lastMousePosition;
+   private Point _lastMousePosition;
     private Point _startMousePosition;
     private Control? _selectedContainer;
 
@@ -32,22 +27,6 @@ public class MapItemsControl : SelectingItemsControl
         SelectionMode = SelectionMode.Multiple;
         SelectionChanged += OnSelectionChanged;
     }
-
-    #region ItemTemplate Property
-
-    public static readonly StyledProperty<IDataTemplate?> AnnotationTemplateProperty =
-        AvaloniaProperty.Register<MapItemsControl, IDataTemplate?>(nameof(AnnotationTemplate));
-
-    [InheritDataTypeFromItems(nameof(ItemsSource))]
-    public IDataTemplate? AnnotationTemplate
-    {
-        get => GetValue(AnnotationTemplateProperty);
-        set => SetValue(AnnotationTemplateProperty, value);
-    }
-
-    #endregion
-
-    #region Selection
 
     private void OnSelectionChanged(object? sender, SelectionChangedEventArgs args)
     {
@@ -68,32 +47,6 @@ public class MapItemsControl : SelectingItemsControl
         }
     }
 
-    public new IList? SelectedItems
-    {
-        get => base.SelectedItems;
-        set => base.SelectedItems = value;
-    }
-
-    /// <inheritdoc />
-    public new ISelectionModel Selection
-    {
-        get => base.Selection;
-        set => base.Selection = value;
-    }
-
-    /// <summary>Gets or sets the selection mode.</summary>
-    /// <remarks>
-    /// Note that the selection mode only applies to selections made via user interaction.
-    /// Multiple selections can be made programmatically regardless of the value of this property.
-    /// </remarks>
-    public new SelectionMode SelectionMode
-    {
-        get => base.SelectionMode;
-        set => base.SelectionMode = value;
-    }
-
-    #endregion
-
     protected override Control CreateContainerForItemOverride(
         object? item,
         int index,
@@ -107,91 +60,6 @@ public class MapItemsControl : SelectingItemsControl
     {
         return NeedsContainer<MapItem>(item, out recycleKey);
     }
-
-    #region Selection rect
-
-    private double _selectionLeft;
-
-    public static readonly DirectProperty<MapItemsControl, double> SelectionLeftProperty =
-        AvaloniaProperty.RegisterDirect<MapItemsControl, double>(
-            nameof(SelectionLeft),
-            o => o.SelectionLeft,
-            (o, v) => o.SelectionLeft = v
-        );
-
-    public double SelectionLeft
-    {
-        get => _selectionLeft;
-        set => SetAndRaise(SelectionLeftProperty, ref _selectionLeft, value);
-    }
-
-    private double _selectionTop;
-
-    public static readonly DirectProperty<MapItemsControl, double> SelectionTopProperty =
-        AvaloniaProperty.RegisterDirect<MapItemsControl, double>(
-            nameof(SelectionTop),
-            o => o.SelectionTop,
-            (o, v) => o.SelectionTop = v
-        );
-
-    public double SelectionTop
-    {
-        get => _selectionTop;
-        set => SetAndRaise(SelectionTopProperty, ref _selectionTop, value);
-    }
-
-    private double _selectionWidth;
-
-    public static readonly DirectProperty<MapItemsControl, double> SelectionWidthProperty =
-        AvaloniaProperty.RegisterDirect<MapItemsControl, double>(
-            nameof(SelectionWidth),
-            o => o.SelectionWidth,
-            (o, v) => o.SelectionWidth = v
-        );
-
-    public double SelectionWidth
-    {
-        get => _selectionWidth;
-        set => SetAndRaise(SelectionWidthProperty, ref _selectionWidth, value);
-    }
-
-    private double _selectionHeight;
-
-    public static readonly DirectProperty<MapItemsControl, double> SelectionHeightProperty =
-        AvaloniaProperty.RegisterDirect<MapItemsControl, double>(
-            nameof(SelectionHeight),
-            o => o.SelectionHeight,
-            (o, v) => o.SelectionHeight = v
-        );
-
-    public double SelectionHeight
-    {
-        get => _selectionHeight;
-        set => SetAndRaise(SelectionHeightProperty, ref _selectionHeight, value);
-    }
-
-    #endregion
-
-    #region Drag state
-
-    private DragState _dragState;
-
-    public static readonly DirectProperty<MapItemsControl, DragState> DragStateProperty =
-        AvaloniaProperty.RegisterDirect<MapItemsControl, DragState>(
-            nameof(DragState),
-            o => o.DragState,
-            (o, v) => o.DragState = v
-        );
-
-    public DragState DragState
-    {
-        get => _dragState;
-        set => SetAndRaise(DragStateProperty, ref _dragState, value);
-    }
-
-    #endregion
-
-    #region Pointer Events
 
     protected override void OnPointerPressed(PointerPressedEventArgs e)
     {
@@ -385,81 +253,4 @@ public class MapItemsControl : SelectingItemsControl
             Zoom = newZoom;
         }
     }
-
-    #endregion
-
-    #region Cursor position
-
-    private GeoPoint _cursorPosition;
-
-    public static readonly DirectProperty<MapItemsControl, GeoPoint> CursorPositionProperty =
-        AvaloniaProperty.RegisterDirect<MapItemsControl, GeoPoint>(
-            nameof(CursorPosition),
-            o => o.CursorPosition,
-            (o, v) => o.CursorPosition = v
-        );
-
-    public GeoPoint CursorPosition
-    {
-        get => _cursorPosition;
-        set => SetAndRaise(CursorPositionProperty, ref _cursorPosition, value);
-    }
-
-    #endregion
-
-    #region CenterMap
-
-    private GeoPoint _centerMap;
-
-    public static readonly DirectProperty<MapItemsControl, GeoPoint> CenterMapProperty =
-        AvaloniaProperty.RegisterDirect<MapItemsControl, GeoPoint>(
-            nameof(CenterMap),
-            o => o.CenterMap,
-            (o, v) => o.CenterMap = v
-        );
-
-    public GeoPoint CenterMap
-    {
-        get => _centerMap;
-        set => SetAndRaise(CenterMapProperty, ref _centerMap, value);
-    }
-
-    #endregion
-
-    #region Provider
-
-    private ITileProvider _provider = EmptyTileProvider.Instance;
-    public static readonly DirectProperty<MapItemsControl, ITileProvider> ProviderProperty =
-        AvaloniaProperty.RegisterDirect<MapItemsControl, ITileProvider>(
-            nameof(Provider),
-            o => o.Provider,
-            (o, v) => o.Provider = v
-        );
-
-    public ITileProvider Provider
-    {
-        get => _provider;
-        set => SetAndRaise(ProviderProperty, ref _provider, value);
-    }
-
-    #endregion
-
-    #region Zoom
-
-    private int _zoom = 8;
-
-    public static readonly DirectProperty<MapItemsControl, int> ZoomProperty =
-        AvaloniaProperty.RegisterDirect<MapItemsControl, int>(
-            nameof(Zoom),
-            o => o.Zoom,
-            (o, v) => o.Zoom = v
-        );
-
-    public int Zoom
-    {
-        get => _zoom;
-        set => SetAndRaise(ZoomProperty, ref _zoom, value);
-    }
-
-    #endregion
 }
