+++ src/Asv.Avalonia/Controls/ContentDialog/ContentDialog.cs
@@ -7,6 +7,7 @@ using Avalonia.Input;
 using Avalonia.Interactivity;
 using Avalonia.Threading;
 using Avalonia.VisualTree;
+using ContentControl = Avalonia.Controls.ContentControl;
 
 namespace Asv.Avalonia;
