+++ src/Asv.Avalonia/Controls/TreePage/TreePageView.axaml.cs
@@ -1,6 +1,4 @@
-﻿using Avalonia;
-using Avalonia.Controls;
-using Avalonia.Markup.Xaml;
+﻿using Avalonia.Controls;
 
 namespace Asv.Avalonia;
