+++ src/Asv.Avalonia/Controls/Workspace/WorkspaceItem.axaml.cs
@@ -1,18 +1,8 @@
-﻿using Avalonia;
-using Avalonia.Controls.Primitives;
+﻿using Avalonia.Controls.Primitives;
 
 namespace Asv.Avalonia;
 
-public class WorkspaceItem : HeaderedContentControl
+public partial class WorkspaceItem : HeaderedContentControl
 {
     public WorkspaceItem() { }
-
-    public static readonly StyledProperty<WorkspaceDock> PositionProperty =
-        AvaloniaProperty.Register<WorkspaceItem, WorkspaceDock>(nameof(Position));
-
-    public WorkspaceDock Position
-    {
-        get => GetValue(PositionProperty);
-        set => SetValue(PositionProperty, value);
-    }
 }
