+++ src/Asv.Avalonia/Controls/Workspace/WorkspaceItem.properties.cs
@@ -1,0 +1,15 @@
+using Avalonia;
+
+namespace Asv.Avalonia;
+
+public partial class WorkspaceItem
+{
+    public static readonly StyledProperty<WorkspaceDock> PositionProperty =
+        AvaloniaProperty.Register<WorkspaceItem, WorkspaceDock>(nameof(Position));
+
+    public WorkspaceDock Position
+    {
+        get => GetValue(PositionProperty);
+        set => SetValue(PositionProperty, value);
+    }
+}
