+++ src/Asv.Avalonia/Shell/Pages/Settings/HotKeys/SettingsKeymapView.axaml.cs
@@ -63,15 +63,4 @@ public partial class SettingsKeymapView : UserControl
 
         vm.SelectedItem.Value.NewHotKeyValue.Value += rawGesture;
     }
-
-    private bool IsModifierKey(Key key)
-    {
-        return key
-            is Key.LeftShift
-                or Key.RightShift
-                or Key.LeftCtrl
-                or Key.RightCtrl
-                or Key.LeftAlt
-                or Key.LeftAlt;
-    }
 }
