+++ src/Asv.Avalonia/Shell/Pages/Settings/HotKeys/SettingsKeymapView.properties.cs
@@ -1,0 +1,18 @@
+using Avalonia.Controls;
+using Avalonia.Input;
+
+namespace Asv.Avalonia;
+
+public partial class SettingsKeymapView
+{
+    private bool IsModifierKey(Key key)
+    {
+        return key
+            is Key.LeftShift
+            or Key.RightShift
+            or Key.LeftCtrl
+            or Key.RightCtrl
+            or Key.LeftAlt
+            or Key.LeftAlt;
+    }
+}
