+++ src/Asv.Avalonia/Shell/View/DockWindow.properties.cs
@@ -1,0 +1,17 @@
+using Avalonia;
+
+namespace Asv.Avalonia;
+
+public partial class DockWindow
+{
+    public static readonly StyledProperty<bool> CloseRequestedProperty = AvaloniaProperty.Register<
+        DockWindow,
+        bool
+    >(nameof(CloseRequested));
+
+    public bool CloseRequested
+    {
+        get => GetValue(CloseRequestedProperty);
+        set => SetValue(CloseRequestedProperty, value);
+    }
+}
